build(all): move outDir/rootDir to tsconfig.build.json (#131)
The two config options are only relevant when emitting js files, which
can only be done with the tsconfig.build.json files, where also the
paths config option is reset. Setting the rootDir to "src" without also
resetting the paths config leads to validation errors in VSCode for the
tsconfig.json files.