fix(integration-tests): do not emit declaration files (to memory-fs)
First of all, we don't need them. And secondly, this reduces the webpack
output when running the tests.

I would have liked to also reduce the stats that are printed, but because
of the following bug, only the one of the two bundled files is printed:
https://github.com/facebook/jest/issues/2441
Jest PR with a fix: https://github.com/facebook/jest/pull/6871

When the PR is merged, we could use this stats config for example:

```json
{
  all: false,
  assets: true,
  colors: true,
  timings: true
}
```