docs(all): improve consistency of main README
The main motivation for the changes were on the one hand the guides on "Writing a Feature App/Service" which should show the final result right at the beginning and on the other hand the example of a Logger Feature Services for the deprecation warnings is not a good example and was therefore removed. We already show elsewhere how a Feature Service dependency can be used.