fix(history-service): navigation after reload (#175)
fixes #171

Instead of using a memory history under the hood for the consumer history,
and storing the entries in session storage, we just compute the consumer
path from the root location, and store all consumer states in the root
history state.

This change also removes the memory history. We don't need a facade for
memory histories, since they don't affect the URL and browser history.
For server-side rendering we will provide a static history instead
(in a separate upcoming PR).