feat(async-ssr-manager): move timeout from config to options
BREAKING CHANGE: The `asyncSsrManagerDefinition` has been replaced with
`defineAsyncSsrManager`. This factory function accepts options to set
the `timeout`, which previously had to be set by the integrator via
Feature Service configs. This is in preparation of an upcoming breaking
change that removes the ability to provide Feature Service configs
through the Feature Service registry.