feat(async-ssr-manager): support concurrent `renderUntilCompleted` calls
When the integrator is not a React app, and renders multiple feature
apps concurrently with separate `ReactDOM.renderToString` calls, each
wrapped with `renderUntilCompleted`, we must scope the scheduled async
operations to the respective `renderUntilCompleted` call. This is
accomplished by leveraging `AsyncLocalStorage`.

Note: This scenario did kind of work already, by accident, up until
v2.9.0. But it was broken with #702 which we shipped with
[v2.10.0](https://github.com/sinnerschrader/feature-hub/releases/tag/v2.10.0).